requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
numpy>=1.19.0
pandas>=1.3.0
scikit-learn>=1.0.0
